fixing broken tests and wacky rebase
