+++ marklogic-data-hub/src/server-side/impl/flow-lib.sjs
@@ -401,6 +401,7 @@ function instanceToCanonicalXml(entityInstance) {
       }
     x.endElement();
   x.endDocument();
+  return x.toNode();
 };
 
 function setDefaultOptions(options, flow) {
